Update 01_todos.spec.js
Updated Assertion on Line 83. If a user was to mark more than 1 item as complete, the original code would fail, as it appears to only count the DIV element.
Should a user mark 2 items as complete, the application only renders a single DIV.
Ive changed the line to allow for multiple items to be checked as completed and perform an assertion based on the panel.
Perhaps this isn't the best locator to use, but I have tested that the numbers within the ID do not increment each time the completed panel is rendered in the DOM.
In cases where an item is marked as completed, then all completed items are deleted, and new items then created and marked as completed.
I could be wrong in my approach, so happy for someone to correct me.